requirements.txt
alabaster==0.7.12
Babel==2.9.1
certifi==2021.5.30
charset-normalizer==2.0.4
docutils==0.16
idna==3.2
imagesize==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
pkg_resources==0.0.0
prettytable==2.1.0
pyfiglet==0.8.post1
Pygments==2.10.0
pyparsing==2.4.7
python-nmap==0.6.4
python3-nmap==1.5.0
pytz==2021.1
requests==2.26.0
scapy==2.4.5
simplejson==3.17.4
snowballstemmer==2.1.0
Sphinx==4.1.2
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
urllib3==1.26.6
wcwidth==0.2.5
